add: aws util functions and export it
