+++ lib/utils/AWS/helper.ts
@@ -1,0 +1,4 @@
+export const region = import.meta.env.VITE_AWS_REGION
+export const bucket = import.meta.env.VITE_S3_BUCKET
+export const accessKeyId = import.meta.env.VITE_AWS_ACCESS_KEY_ID
+export const secretAccessKey = import.meta.env.VITE_AWS_SECRET_ACCESS_KEY
